Name change for the project and a cleared README file
The original name for this project was Wrengo, given that it is essentially
wren with Go bindings, a wrapper and a somewhat friendly API to be developed
on. Since then I've been thinking of ways to use this effectivly and it'll
start off by integrating things that allow it to asist with the build of a
project(s).

It also needs some basic classes that work well with the various types,
strings etc... Part of this work is underway, however tests need to be
written before much more can be done in terms of what is exposed as part
of the public API for si.

Also si is always lower-case.