Starting to nativly implements tests
Not much here at the moment, just testing out a way of being able to make
assertions as part of the language itself. Seems to work so far although there
could be improvmenets made to how the tests are written. Each method currently
has to be staticly called.

This is fine for now.